# movies_auth/src/openapi.yaml
openapi: '3.0.0'
info:
  title: Auth OpenAPI
  description: OpenAPI schema for authenticate service of online movies
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: http://{host}:{port}/api/{version}/auth
    description: Auth server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '5000'
          - '80'
        default: '5000'
      version:
        default: v1
paths:
...
  /role:
    post:
      description: Returns created role
      requestBody:
        description: role request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '200':
          description: created role
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
  /role/{role_id}:
    put:
      description: Returns changed role
      requestBody:
        description: role request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
      responses:
        '200':
          description: changed role
          content:
            'application/json':
              schema:
                type: object
                required:
                  - item
                  - message
                  - properties
                properties:
                  item:
                    $ref: '#/components/schemas/Role'
                  message:
                    type: string
                  properties:
                    type: object
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
      - name: role_id
        in: path
        description: role uuid
        required: true
        schema:
          type: string
...
  /user:
    post:
      description: Returns created user
      requestBody:
        description: user request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                email:
                  type: string
      responses:
        '200':
          description: created user
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    put:
      description: Returns changed user
      requestBody:
        description: user request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                email:
                  type: string
      responses:
        '200':
          description: changed user
          content:
            'application/json':
              schema:
                type: object
                required:
                  - item
                  - message
                  - properties
                properties:
                  item:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                  properties:
                    type: object
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
        - name: Authorization
          in: header
          description: JWT token
          required: true
          schema:
            type: string
  /user_role_add:
    post:
      description: Assigns role to user
      requestBody:
        description: user role request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - user_id
                - role_id
              properties:
                user_id:
                  type: string
                role_id:
                  type: string
      responses:
        '200':
          description: role assigned to user
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/UserRole'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
  /user_role_revoke:
    post:
      description: Revoke role from user
      requestBody:
        description: user role request body
        required: true
        content:
          'application/json':
            schema:
              type: object
              required:
                - user_id
                - role_id
              properties:
                user_id:
                  type: string
                role_id:
                  type: string
      responses:
        '200':
          description: role revoked from user
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/UserRole'
        '401':
          description: Unauthorized
        default:
          description: error payload
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Error'
    parameters:
      - name: Authorization
        in: header
        description: JWT token
        required: true
        schema:
          type: string
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Role:
      type: object
      required:
        - uuid
        - name
      properties:
        uuid:
          type: string
        name:
          type: string
...
    User:
      type: object
      required:
        - uuid
        - username
        - password
      properties:
        uuid:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
    UserRole:
      type: object
      required:
        - uuid
        - user_id
        - role_id
      properties:
        uuid:
          type: string
        user_id:
          type: string
        role_id:
          type: string